backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.2
slowapi==0.1.9
pytest==8.2.0
pytest-asyncio==0.23.6
